move operations from updater root to appropriate file
