+++ stats/stats/src/charts/db_interaction/updater/common_operations/mod.rs
@@ -1,0 +1,4 @@
+//! Collection of common operations to perform while updating.
+//! Can be useful for any chart regardless of their type.
+
+mod read;
